# src/community/ogcapi/ogcapi-coverages/src/main/resources/org/geoserver/ogcapi/v1/coverages/openapi.yaml
openapi: 3.0.2
info:
  title: A sample API conforming to the OGC API - Coverages standard
  description: 'This is a sample OpenAPI definition that conforms to the OGC API -
    Coverages Service specification (conformance classes: "Core", "Subset", "JSON",
    "HTML" and "OpenAPI 3.0").'
  contact:
    name: Acme Corporation
    url: http://example.org/
    email: info@example.org
  license:
    name: CC-BY 4.0 license
    url: https://creativecommons.org/licenses/by/4.0/
  version: 1.0.0
servers:
  - url: https://virtserver.swaggerhub.com/constantinius/ogc-api-coverages/1.0.0
    description: SwaggerHub API Auto Mocking
  - url: https://dev.example.org/
    description: Development server
  - url: https://data.example.org/
    description: Production server
...
paths:
  /:
    get:
      tags:
        - Capabilities
      summary: landing page of this API
      description: The landing page provides links to the API definition, the Conformance
        statements and the metadata about the coverage data in this dataset.
      operationId: getLandingPage
      parameters:
        - $ref: '#/components/parameters/f'
      responses:
        "200":
          description: links to the API capabilities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/root'
            text/html:
              schema:
                type: string
...
  /conformance:
    get:
      tags:
        - Capabilities
      operationId: getConformanceDeclaration
      summary: API conformance definition
      description: list of all conformance classes that the server conforms to
      parameters:
        - $ref: '#/components/parameters/f'
      responses:
        "200":
          description: the URIs of all requirements classes supported by the server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/req-classes'
        default:
          description: An error occured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
  /collections:
    get:
      tags:
        - Capabilities
      summary: describe the collections in the dataset
      operationId: getCollections
      parameters:
        - $ref: '#/components/parameters/f'
      responses:
        "200":
          description: Metadata about the collections shared by this API.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/content'
            text/html:
              schema:
                type: string
        default:
          description: An error occured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
  /collections/{collectionId}:
    get:
      tags:
        - Capabilities
      summary: Collection Information is the set of metadata which describes a single
        collection, or in the the case of API-Coverages, a single Coverage.
      operationId: describeCollection
      parameters:
        - $ref: '#/components/parameters/collectionId'
        - $ref: '#/components/parameters/f'
      responses:
        "200":
          description: Metadata about the {collectionId} collection shared by this API.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/collectionInfo'
            text/html:
              schema:
                type: string
        default:
          description: An error occured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/exception'
            text/html:
              schema:
                type: string
...
components:
  schemas:
    exception:
      required:
        - code
      type: object
      properties:
        code:
          type: string
        description:
          type: string
    root:
      required:
        - links
      type: object
      properties:
        title:
          type: string
          description: The title of the API
        description:
          type: string
          description: A textual description of the API
        links:
          type: array
          description: Links to the resources exposed through this API.
          example:
            - href: http://data.example.org/
              rel: self
              type: application/json
              title: this document
            - href: http://data.example.org/api
              rel: service-desc
              type: application/openapi+json;version=3.0
              title: the API definition
            - href: http://data.example.org/conformance
              rel: conformance
              type: application/json
              title: OGC conformance classes implemented by this API
            - href: http://data.example.org/collections
              title: Metadata about the resource collections
          items:
            $ref: '#/components/schemas/link'
    req-classes:
      required:
        - conformsTo
      type: object
      properties:
        conformsTo:
          type: array
          example:
            - http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/core
            - http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/collections
            - http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/oas3
            - http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/html
            - http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/geojson
            - http://www.opengis.net/spec/ogcapi-coverages-1/1.0/conf/core
          items:
            type: string
    link:
      required:
        - href
      type: object
      properties:
        href:
          type: string
          example: http://data.example.com/collections/dem/coverage
        rel:
          type: string
          example: alternate
        type:
          type: string
          example: application/json
        hreflang:
          type: string
          example: en
        title:
          type: string
          example: Digital Elevation Model
    content:
      required:
        - collections
        - links
      type: object
      properties:
        collections:
          type: array
          example:
            name: dem
            title: Digital Elevation Model
            description: Digital Elevation Model
            links:
              - href: http://data.example.org/collections/dem?f=json
                rel: self
                type: application/json
                title: Digital Elevation Model
              - href: http://data.example.org/collections/dem?f=html
                rel: alternate
                type: application/json
                title: Digital Elevation Model
              - href: http://data.example.org/collections/dem/coverage
                rel: coverage
                type: application/json
                title: Digital Elevation Model
            extent:
              crs: http://www.opengis.net/def/crs/OGC/1.3/CRS84
              spatial:
                - -180
                - -90
                - 180
                - 90
              trs: http://www.opengis.net/def/uom/ISO-8601/0/Gregorian
              temporal:
                - 2020-01-06T14:00:00Z
                - 2020-02-07T22:00:00Z
            crs:
              - http://www.opengis.net/def/crs/OGC/1.3/CRS84
          items:
            $ref: '#/components/schemas/collectionInfo'
        links:
          type: array
          example:
            - href: http://data.example.org/collections?f=json
              rel: self
              type: application/json
              title: this document
            - href: http://data.example.org/collections?f=html
              rel: alternate
              type: text/html
              title: this document as HTML
          items:
            $ref: '#/components/schemas/link'
    collectionInfo:
      required:
        - links
        - name
      type: object
      properties:
        name:
          type: string
          description: identifier of the collection used, for example, in URIs
          example: dem
        title:
          type: string
          description: human readable title of the collection
          example: Digital Elevation Model
        description:
          type: string
          description: a description of the data in the collection
          example: Digital Elevation Model
        links:
          type: array
          example:
            - href: http://data.example.org/collections/dem?f=json
              rel: self
              type: application/json
              title: Digital Elevation Model
            - href: http://data.example.org/collections/dem?f=html
              rel: alternate
              type: application/json
              title: Digital Elevation Model
            - href: http://data.example.org/collections/dem/coverage
              rel: coverage
              type: application/json
              title: Digital Elevation Model
            - href: http://data.example.org/collections/dem/coverage/domainset
              rel: domainset
              type: application/json
              title: Digital Elevation Model
            - href: http://data.example.org/collections/dem/coverage/rangetype
              rel: rangetype
              type: application/json
              title: Digital Elevation Model
            - href: http://data.example.org/collections/dem/coverage/metadata
              rel: metadata
              type: application/json
              title: Digital Elevation Model
            - href: http://data.example.org/collections/dem/coverage/rangeset
              rel: rangeset
              type: application/json
              title: Digital Elevation Model
            - href: http://data.example.org/collections/dem/coverage/description
              rel: description
              type: application/json
              title: Digital Elevation Model
            - href: http://data.example.org/collections/dem/coverage/all
              rel: data
              type: application/json
              title: Digital Elevation Model
          items:
            $ref: '#/components/schemas/link'
        extent:
          $ref: '#/components/schemas/extent'
        crs:
          type: array
          description: The coordinate reference systems in which geometries may be
            retrieved. Coordinate reference systems are identified by a URI. The first
            coordinate reference system is the coordinate reference system that is
            used by default. This is always "http://www.opengis.net/def/crs/OGC/1.3/CRS84",
            i.e. WGS84 longitude/latitude.
          example:
            - http://www.opengis.net/def/crs/OGC/1.3/CRS84
          items:
            type: string
          default:
            - http://www.opengis.net/def/crs/OGC/1.3/CRS84
    extent:
      type: object
      properties:
        crs:
          type: string
          description: Coordinate reference system of the coordinates in the spatial
            extent (property `spatial`). In the Core, only WGS84 longitude/latitude
            is supported. Extensions may support additional coordinate reference systems.
          default: http://www.opengis.net/def/crs/OGC/1.3/CRS84
          enum:
            - http://www.opengis.net/def/crs/OGC/1.3/CRS84
        spatial:
          maxItems: 6
          minItems: 4
          type: array
          description: West, north, east, south edges of the spatial extent. The minimum
            and maximum values apply to the coordinate reference system WGS84 longitude/latitude
            that is supported in the Core. If, for example, a projected coordinate
            reference system is used, the minimum and maximum values need to be adjusted.
          example:
            - -180
            - -90
            - 180
            - 90
          items:
            type: number
        trs:
          type: string
          description: Temporal reference system of the coordinates in the temporal
            extent (property `temporal`). In the Core, only the Gregorian calendar
            is supported. Extensions may support additional temporal reference systems.
          default: http://www.opengis.net/def/uom/ISO-8601/0/Gregorian
          enum:
            - http://www.opengis.net/def/uom/ISO-8601/0/Gregorian
        temporal:
          maxItems: 2
          minItems: 2
          type: array
          description: Begin and end times of the temporal extent.
          example:
            - 2020-01-06T14:00:00Z
            - 2020-02-07T22:00:00Z
          items:
            type: string
            format: dateTime
...
  parameters:
...
    f:
      name: f
      in: query
      description: |
        The optional f parameter indicates the output format which the server shall provide as part of the response document. It has preference over the HTTP Accept header. The default format is JSON.
      required: false
      style: form
      explode: false
      schema:
        type: string
        default: json
        enum:
          - json
          - html
          - yaml
    collectionId:
      name: collectionId
      in: path
      description: Identifier (name) of a specific collection
      required: true
      style: simple
      explode: false
      schema:
        type: string
...